Change ringbuffer_getchar to return an int so that we can indicate emptiness with -1, and have an 8-bit clean stream.

+++ ringbuffer.h
@@ -27,7 +27,7 @@ uint32_t ringbuffer_used(RINGBUF*);
 uint32_t ringbuffer_free(RINGBUF*);
 
 int32_t ringbuffer_addchar(RINGBUF*, uint8_t);
-uint8_t ringbuffer_getchar(RINGBUF*); 
+int ringbuffer_getchar(RINGBUF*); 
 
 uint8_t *ringbuffer_getbulkpointer(RINGBUF*); 
 int32_t ringbuffer_getbulkcount(RINGBUF*);
